ref: Move PromiseBuffer and Error to utils package
