+++ packages/core/src/basebackend.ts
@@ -1,8 +1,8 @@
 import { Scope } from '@sentry/hub';
 import { Breadcrumb, SentryEvent, SentryEventHint, SentryResponse, Severity, Transport } from '@sentry/types';
+import { SentryError } from '@sentry/utils/error';
 import { logger } from '@sentry/utils/logger';
 import { serialize } from '@sentry/utils/object';
-import { SentryError } from './error';
 import { Backend, Options } from './interfaces';
 import { NoopTransport } from './transports/noop';
