+++ packages/core/src/index.ts
@@ -11,8 +11,6 @@ export { API } from './api';
 export { BaseClient } from './baseclient';
 export { BackendClass, BaseBackend } from './basebackend';
 export { Dsn } from './dsn';
-export { SentryError } from './error';
-export { PromiseBuffer } from './promisebuffer';
 export { Backend, Client, LogLevel, Options } from './interfaces';
 export { initAndBind, ClientClass } from './sdk';
 export { NoopTransport } from './transports/noop';
