+++ packages/node/src/transports/http.ts
@@ -1,5 +1,5 @@
-import { SentryError } from '@sentry/core';
 import { SentryResponse, TransportOptions } from '@sentry/types';
+import { SentryError } from '@sentry/utils/error';
 import * as http from 'http';
 import * as HttpsProxyAgent from 'https-proxy-agent';
 import { BaseTransport } from './base';
